pyproject.toml
[tool.poetry]
name = "pcnewsgpt"
version = "0.1.0"
description = "KI-Abfrage der PC-News Inhalte über RAG"
authors = ["Andreas Kunar"]
license = "Apache"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "0.0.177"
chromadb = "0.3.23"
urllib3 = "2.0.2"
pdfminer-six = "20221105"
python-dotenv = "1.0.0"
unstructured = "0.6.6"
tqdm = "4.65.0"
regex = "^2023.8.8"
importlib = "^1.0.4"
llama-cpp-python = "^0.1.83"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
